"render" helper function, method name formatter, save all HTML
* Added ich.render() as a helper alias to rendering a template by
calling its method in case the template name was not a valid method
name, so calling it would not look as weird (e.g. ich["tmpl-name"](data,
raw)). * Created ich.formatName() to create a camel-case name from a
file path or URL, removing the extension of a filename that may have
been identified.  This is used internally to give a template a name when
no name is specified. * Fixed ich.loadTemplates() to add any remaining
HTML to a template after all sub-templates and partials have been
extracted.